+++ packages/sector/tool/dartdoc.dart
@@ -3,8 +3,7 @@
 import 'dart:io' as io;
 
 import 'package:args/args.dart';
-import 'package:shelf/shelf_io.dart' as shelf_io;
-import 'package:shelf_static/shelf_static.dart';
+import 'package:oath/tools/dartdoc.dart';
 
 /// Generates dartdoc for the project.
 void main(List<String> args) async {
@@ -17,16 +16,10 @@ void main(List<String> args) async {
       negatable: false,
     )
     ..addFlag(
-      'generate',
-      help: 'Generates the documentation.',
-      defaultsTo: true,
-    )
-    ..addOption(
       'preview',
       abbr: 'p',
-      help: 'Opens the generated documentation in a browser.',
-      allowed: ['none', 'listen', 'listen-open-browser'],
-      defaultsTo: isCI ? 'none' : 'listen-open-browser',
+      help: 'Serve the documentation to preview.',
+      defaultsTo: !isCI,
     );
   final results = parser.parse(args);
 
@@ -35,46 +28,7 @@ void main(List<String> args) async {
     return;
   }
 
-  final preview = results['preview'] as String;
-  final generate = results['generate'] as bool;
-
-  if (generate) {
-    final dartdoc = await io.Process.start('dart', ['doc']);
-
-    dartdoc.stdout.listen(io.stdout.add);
-    dartdoc.stderr.listen(io.stderr.add);
-
-    final exitCode = await dartdoc.exitCode;
-    if (exitCode != 0) {
-      throw Exception('Failed to generate documentation: $exitCode');
-    }
-  }
-
-  if (preview != 'none') {
-    final handler = createStaticHandler('doc/api');
-    final server = await shelf_io.serve(handler, 'localhost', 8080);
-
-    io.stdout.writeln(
-      'Serving documentation on http://localhost:8080/index.html',
-    );
-    io.stdout.writeln('Press Ctrl+C to exit.');
-
-    if (preview == 'listen-open-browser') {
-      final process = await io.Process.start(
-        'open',
-        ['http://localhost:8080/index.html'],
-      );
-
-      process.stdout.listen(io.stdout.add);
-      process.stderr.listen(io.stderr.add);
-
-      final exitCode = await process.exitCode;
-      if (exitCode != 0) {
-        throw Exception('Failed to open browser: $exitCode');
-      }
-    }
-
-    await io.ProcessSignal.sigint.watch().first;
-    await server.close(force: true);
-  }
+  await runDartdoc(
+    mode: results.flag('preview') ? DartdocMode.preview : DartdocMode.generate,
+  );
 }